% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/effects.R
\name{effect.cond}
\alias{effect.cond}
\title{Calculate the average partial effects on conditional expected values}
\usage{
effect.cond(model)
}
\arguments{
\item{model}{The AER model to use}
}
\value{
- call
- coefficients
- tableAPEcond
}
\description{
This function calculates the different
fitted values from a probit model estimated
with AER package
}
\examples{
library(wooldridge)
data(mroz)
model <- tobit(hours ~ nwifeinc + educ, data=mroz)
effect.cond(model)

}
