% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fitting.R
\name{fitted.tobit}
\alias{fitted.tobit}
\title{Fitting values for a tobit model}
\usage{
\method{fitted}{tobit}(model)
}
\arguments{
\item{model}{The model to use}
}
\value{
res: A matrix of the infile
}
\description{
This function calculates the different
fitted values from a probit model estimated
with AER package PyPos, EyPos, and Ey
}
