% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fitting.R
\name{fitTobit}
\alias{fitTobit}
\title{Fitting values for a tobit model}
\usage{
fitTobit(model)
}
\arguments{
\item{model}{The model to use}
}
\value{
Returns a tobitExpected object with two values:

+ call: the tobit Call
+ table: A 4 columns matrix observed value, PyPos, EyPos, and Ey
}
\description{
This function calculates the different
fitted values from a probit model estimated
with AER package
}
\examples{
library(wooldridge)
data(mroz)
model <- tobit(hours ~ nwifeinc + educ, data=mroz)
fitTobit(model)

}
