% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/effects.R
\name{se.effect.uncond}
\alias{se.effect.uncond}
\title{Calculates the standard errors of the
partial effects on unconditional expected values
It requires a tobit model estimated with the  AER package}
\usage{
se.effect.uncond(model, R, data)
}
\arguments{
\item{model}{+ model: The AER model to use
+ R: the number of repetitions
+ data: the data frame that was used for the model}
}
\value{
- call: the call of the model
- table: A 4 x k table gathering the mean of
the average partial effets
- R: the number of repetitions
}
\description{
Calculates the standard errors of the
partial effects on unconditional expected values
It requires a tobit model estimated with the  AER package
}
\examples{
library(wooldridge)
data(mroz)
model <- tobit(hours ~ nwifeinc + educ, data=mroz)
sem <- se.effect.uncond(model)
sem  #raw output
summary.uncond.eff(sem)  #output the formatted table

}
