% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/effects.R
\name{varClass}
\alias{varClass}
\title{Extract the classes of the model independent variables
Class them into numeric and factor variables}
\usage{
varClass(model)
}
\description{
Extract the classes of the model independent variables
Class them into numeric and factor variables
}
